+++ lib/es6/belt_internalSetBuckets.js
@@ -3,18 +3,6 @@
 import * as Curry from "./curry.js";
 import * as Belt_Array from "./belt_Array.js";
 
-function copyBucket(c) {
-  if (c === undefined) {
-    return c;
-  }
-  let head = {
-    key: c.key,
-    next: undefined
-  };
-  copyAuxCont(c.next, head);
-  return head;
-}
-
 function copyAuxCont(_c, _prec) {
   while(true) {
     let prec = _prec;
@@ -31,6 +19,18 @@ function copyAuxCont(_c, _prec) {
     _c = c.next;
     continue;
   };
+}
+
+function copyBucket(c) {
+  if (c === undefined) {
+    return c;
+  }
+  let head = {
+    key: c.key,
+    next: undefined
+  };
+  copyAuxCont(c.next, head);
+  return head;
 }
 
 function copyBuckets(buckets) {
