+++ lib/es6/js.js
@@ -3,8 +3,6 @@
 
 let Internal = {};
 
-let MapperRt;
-
 let Null;
 
 let Undefined;
@@ -76,7 +74,6 @@ let $$Map;
 let $$WeakMap;
 
 export {
-  MapperRt,
   Internal,
   Null,
   Undefined,
