+++ lib/js/belt_internalBuckets.js
@@ -4,6 +4,19 @@ let Curry = require("./curry.js");
 let Belt_Array = require("./belt_Array.js");
 let Caml_option = require("./caml_option.js");
 
+function copyBucket(c) {
+  if (c === undefined) {
+    return c;
+  }
+  let head = {
+    key: c.key,
+    value: c.value,
+    next: undefined
+  };
+  copyAuxCont(c.next, head);
+  return head;
+}
+
 function copyAuxCont(_c, _prec) {
   while(true) {
     let prec = _prec;
@@ -21,19 +34,6 @@ function copyAuxCont(_c, _prec) {
     _c = c.next;
     continue;
   };
-}
-
-function copyBucket(c) {
-  if (c === undefined) {
-    return c;
-  }
-  let head = {
-    key: c.key,
-    value: c.value,
-    next: undefined
-  };
-  copyAuxCont(c.next, head);
-  return head;
 }
 
 function copyBuckets(buckets) {
