+++ lib/js/js.js
@@ -3,8 +3,6 @@
 
 let Internal = {};
 
-let MapperRt;
-
 let Null;
 
 let Undefined;
@@ -75,7 +73,6 @@ let $$Map;
 
 let $$WeakMap;
 
-exports.MapperRt = MapperRt;
 exports.Internal = Internal;
 exports.Null = Null;
 exports.Undefined = Undefined;
